% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_w_func.R
\name{wrap_w_func}
\alias{wrap_w_func}
\title{Wrap selected text in source editor with a function}
\usage{
wrap_w_func()
}
\description{
This is the one and only function for now. This function takes
no argument; when run, it will prompt for user input, and then
wrap the selected text in source editor with the function.
Obviously, only work in Rstudio.
}
